requirements.txt
psycopg2-binary==2.9.9
tweepy==4.14.0
python-dotenv==1.0.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.0
email-validator==2.1.0